Add big decimal

+++ SalesTaxes/SalesTaxes/domain/src/main/java/it/tommasoresti/salestaxes/domain/ArticleFactory.java
@@ -1,7 +1,9 @@
 package it.tommasoresti.salestaxes.domain;
 
+import java.math.BigDecimal;
+
 import it.tommasoresti.salestaxes.domain.article.Article;
 
 public interface ArticleFactory {
-    Article make(String description, float price);
+    Article make(String description, BigDecimal price);
 }
